+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 - Added .displayName to components
 
-- Calendar: selectedDates + more tests
+Tests:
+
+- Calendar, TrapFocus, Avatar, Autocomplete, Image, useOnClickOutside, useHandlerRef
+
+- Calendar: selectedDates
 - Flyout: fixed onClickOutside to correctly track clicked components that get unmounted (calendar)
 - useOnClickOutside: new hook
 - useHandlerRef: new hook
 - TrapFocus: new utility, new trap mode
 - Icon: Added vertical alignment in case svg content is smaller that the icon size
-- Avatar: Image attributes, relaxed types, more tests
+- Avatar: Image attributes, relaxed types
 - Avatar: renderImage
 - Image: renderImage
+- Autocomplete: keep focus on the input when clicking on disabled items
+- DropdownMenu: items support passing className now